Handle GKE Connect Gateway URL differences
This PR changes the Connect Gateway builder to always lookup the given membership and determine if the resulting cluster is a GKE cluster. If it is a GKE cluster, it uses the /gkeMemberships collection endpoint. Otherwise, it uses the standard /memberships collection endpiont.

It also drops the e2e test in favor of just using kubectl directly.